+++ src/build.task.ts
@@ -13,9 +13,13 @@ export const task: Task = {
 		const inputFiles = await resolveInputFiles(args._);
 		log.info('inputFiles', inputFiles);
 
-		const dev: boolean = process.env.NODE_ENV !== 'production';
+		// TODO what's the best way to define these types? make `Task` generic? schema validation?
+		const dev: boolean = 'dev' in args ? !!args.dev : process.env.NODE_ENV !== 'production';
 		const watch: boolean = (args.watch as any) || false;
 		const outputDir: string = (args.outputDir as any) || DEFAULT_OUTPUT_DIR;
+		const mapInputOptions = args.mapInputOptions as any;
+		const mapOutputOptions = args.mapOutputOptions as any;
+		const mapWatchOptions = args.mapWatchOptions as any;
 
 		if (inputFiles.length) {
 			const build = createBuild({
@@ -23,6 +27,9 @@ export const task: Task = {
 				inputFiles,
 				outputDir,
 				watch,
+				mapInputOptions,
+				mapOutputOptions,
+				mapWatchOptions,
 			});
 			await build.promise;
 		} else {